feat : [TSK-32] 유저 드롭다운 제거(로그아웃 버튼 설정으로 이동) 및 카드 컴포넌트 모듈화
