+++ app/src/App.tsx
@@ -10,7 +10,7 @@ import Login from './pages/Login';
 import Settings from './pages/Settings';
 import NoDataView from './pages/NoDataView';
 import Onboarding from './pages/Onboarding';
-import UserDropdown from './widgets/UserDropdown';
+import Card from './components/Card';
 import { useTodayFlashcards } from './hooks/useTodayFlashcards';
 import { useNavigationStore } from './stores/navigationStore';
 
@@ -90,43 +90,26 @@ const App: React.FC = () => {
   // 로딩 중 (인증, 온보딩 확인, 데이터 로딩)
   if (authLoading || !onboardingChecked || loading) {
     return (
-      <div style={{ 
-        display: 'flex', 
-        justifyContent: 'center', 
-        alignItems: 'center', 
-        height: '100vh',
-        fontSize: '1.2rem',
-        background: 'linear-gradient(135deg, #667eea 0%, #764ba2 100%)',
-        color: 'white'
-      }}>
+      <Card>
         <div style={{
-          textAlign: 'center',
-          background: 'rgba(255, 255, 255, 0.1)',
-          padding: '40px',
-          borderRadius: '20px',
-          backdropFilter: 'blur(10px)',
-          border: '1px solid rgba(255, 255, 255, 0.2)'
-        }}>
-          <div style={{
-            width: '50px',
-            height: '50px',
-            border: '3px solid rgba(255, 255, 255, 0.3)',
-            borderTop: '3px solid white',
-            borderRadius: '50%',
-            animation: 'spin 1s linear infinite',
-            margin: '0 auto 20px'
-          }}></div>
-          <h2 style={{ marginBottom: '10px' }}>📚 플래시카드 준비 중</h2>
-          <p>GitHub에서 최근 커밋을 분석하고 있습니다...</p>
-          <p style={{ marginTop: '10px', fontSize: '0.9rem', opacity: '0.8' }}>⏱️ 데이터 양에 따라 시간이 조금 걸릴 수 있습니다</p>
-        </div>
+          width: '50px',
+          height: '50px',
+          border: '3px solid rgba(255, 255, 255, 0.3)',
+          borderTop: '3px solid white',
+          borderRadius: '50%',
+          animation: 'spin 1s linear infinite',
+          margin: '0 auto 20px'
+        }}></div>
+        <h2 style={{ marginBottom: '10px', fontSize: '1.25rem' }}>📚 플래시카드 준비 중</h2>
+        <p style={{ fontSize: '1rem' }}>GitHub에서 최근 커밋을 분석하고 있습니다...</p>
+        <p style={{ marginTop: '10px', fontSize: '0.9rem', opacity: '0.8' }}>⏱️ 데이터 양에 따라 시간이 조금 걸릴 수 있습니다</p>
         <style>{`
           @keyframes spin {
             0% { transform: rotate(0deg); }
             100% { transform: rotate(360deg); }
           }
         `}</style>
-      </div>
+      </Card>
     );
   }
 
@@ -207,10 +190,36 @@ const App: React.FC = () => {
           )}
         </div>
         
-        <UserDropdown 
-          user={user}
-          onNavigateToSettings={navigateToSettings}
-        />
+        {currentPage === 'flashcard' && (
+          <button
+            onClick={navigateToSettings}
+            style={{
+              padding: '8px 16px',
+              background: 'rgba(255, 255, 255, 0.95)',
+              color: '#333',
+              border: '1px solid rgba(0, 0, 0, 0.1)',
+              borderRadius: '8px',
+              cursor: 'pointer',
+              fontSize: '1.2rem',
+              transition: 'all 0.2s',
+              boxShadow: '0 2px 8px rgba(0,0,0,0.15)',
+              backdropFilter: 'blur(10px)',
+            }}
+            onMouseEnter={(e) => {
+              e.currentTarget.style.background = 'rgba(255, 255, 255, 1)';
+              e.currentTarget.style.transform = 'translateY(-1px)';
+              e.currentTarget.style.boxShadow = '0 4px 12px rgba(0,0,0,0.2)';
+            }}
+            onMouseLeave={(e) => {
+              e.currentTarget.style.background = 'rgba(255, 255, 255, 0.95)';
+              e.currentTarget.style.transform = 'translateY(0)';
+              e.currentTarget.style.boxShadow = '0 2px 8px rgba(0,0,0,0.15)';
+            }}
+            title="설정"
+          >
+            ⚙️
+          </button>
+        )}
       </nav>
 
       {/* 페이지 컨텐츠 */}
