+++ app/src/pages/NoDataView.css
@@ -1,26 +1,3 @@
-.nodata-container {
-  display: flex;
-  flex-direction: column;
-  height: 100vh;
-  align-items: center;
-  justify-content: center;
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  color: white;
-  text-align: center;
-  padding: 16px;
-}
-
-.nodata-card {
-  background: rgba(255, 255, 255, 0.1);
-  border-radius: 16px;
-  padding: 24px;
-  backdrop-filter: blur(10px);
-  border: 1px solid rgba(255, 255, 255, 0.2);
-  box-shadow: 0 8px 32px rgba(0, 0, 0, 0.1);
-  max-width: 400px;
-  width: 100%;
-}
-
 .nodata-image {
   width: 140px;
   height: auto;
@@ -36,14 +13,14 @@
 }
 
 .nodata-description {
-  font-size: 14px;
+  font-size: 13px;
   margin-bottom: 6px;
   opacity: 0.9;
   line-height: 1.4;
 }
 
 .nodata-hint {
-  font-size: 12px;
+  font-size: 11px;
   margin-bottom: 24px;
   opacity: 0.7;
   line-height: 1.4;
